Began implimenting an LEDManager class
The class will be responsible for taking in HTTP request information, decipering the instructions, and adjusting the LEDs so these processes aren't jumbled into the main .ino file. This promotes abstraction in the code and will provide future flexibility in how we can implement new features into LED modification in the lamp.